# Add a rosbag tool 'transform' to split, merge and transform bag files